[REFACTOR] auth 패키지 프레젠테이션 계층 구현
