+++ src/main/kotlin/io/gijung/aichatbot/auth/presentation/web/response/LoginResponse.kt
@@ -1,0 +1,5 @@
+package io.gijung.aichatbot.auth.presentation.web.response
+
+data class LoginResponse(
+    val token: String,
+)
