--- src/main/kotlin/io/gijung/aichatbot/controller/response/LoginResponse.kt
@@ -1,5 +1,0 @@
-package io.gijung.aichatbot.controller.response
-
-data class LoginResponse(
-    val token: String,
-)
