--- src/main/kotlin/io/gijung/aichatbot/controller/response/SignUpResponse.kt
@@ -1,7 +1,0 @@
-package io.gijung.aichatbot.controller.response
-
-data class SignUpResponse(
-    val id: String,
-    val email: String,
-    val nickname: String
-)
